refactore planetable
